lmkd: Honor lowmem_min_oom_score of 1001 per documentation
In general, when the user sets ro.lmk.lowmem_min_oom_score to 1001,
we won't kill just because the watermark was breached.  The
exception is when "critical_stall" is true, and we would end up
killing anyway.  We change the logic to get rid of this exception,
so the behavior is always as documented.

Furthermore, our change in the logic allows us to avoid two
"psi_parse[...]()" calls in this case.  The results of those
reads of procfs were being ignored in this case, so this change
avoids that waste.

Bug: 406250396
Test: TreeHugger
Change-Id: I556b6f57b193838ad722f87bc7829bf123a5ed5f