lmkd: Make kill_timout_ms of 0 be consistent
In two parts of the code (__mp_event_psi() and mainloop()), if
ro.lmk.kill_timeout_ms was set to 0, we treated this as an
infinite timeout for the kill.

But in mp_event_common(), we treated 0 to mean zero timeout and
that we were ready to instantly kill again.  We change the behavior
in mp_event_common() to also treat this as an infinite timeout.

We update the README to note this (now consistent) behavoir.

Bug: 406071563
Test: TreeHugger
Change-Id: If21657d9537d38ff51f7e6e9ebd3eeefbb1b0d18