Run set_process_group_and_prio asynchronously
The LMKD reaper threads have been observed to become blocked on the
heavily-contended cgroup mutex due to cpuset migration for the victim,
preventing memory reclaim in process_madvise. This delays memory reclaim
that could otherwise occur, resulting in more LMKs than necessary if
reapers have to wait to obtain the cgroup mutex.

Move the code that changes the priority of the victim process to its own
thread so that reapers can begin to reclaim immediately after killing.
Since cpuset migration is serialized in the kernel on the cgroup mutex,
there is not much benefit to having more than a single thread which
calls SetTaskProfiles that does the cpuset migration.

Test: atest lmkd_tests
Test: atest lmkd_unit_test
Change-Id: Ide8ab8833edc6da6589d658a498255178022028c