Add new field for_lmkd_only to struct lmk_procprio
The new field will be used by system_server to send an oom_score_adj
update that should only be applied to lmkd and not kernel. This is
useful for communicating a tweaked oom_score_adj value for a process
whose zram was written back. This kind of process has ~zero value in
terms of memory reclaim so we want lmkd to rank them lower in kill
decisions. For kernel oom killer though, we don't want to influent the
kill decision there. This acts as the last line of defense in case we
got writeback accounting wrong

Bug: 450670546
Test: atest CachedAppOptimizerTest
Flag: com.android.server.am.enable_zram_writeback
Change-Id: I2b2d60d296170625bdd3c55e02fe4ea79d6879da